+++ NEWS.md
@@ -7,7 +7,7 @@ previous versions.  If critical bugs are found in any of the software,
 notice of such bugs and the versions in which they were fixed will be
 noted here, as well.
 
-# perl-Amazon-S3 0.55 (2022-07-21)
+# perl-Amazon-S3 0.55 (2022-07-22))
 
 ## Enhancements
 
@@ -16,7 +16,9 @@ noted here, as well.
 * buckets objects carry region attribute to facilitate signing
 * multipart upload unit tests
 * github workflow build
-
+* revert to using XML::Simple which introduces less dependencies than
+  XML::LibXML and XML::LibXML::Simple
+  
 ## Fixes
 
 * error handling in `complete_multipart_upload()`